Refactoring Image (cont.)
* Shoes3/shoes3#333
